Create Focusable abstract class and refactor Actor and Attribute
Introduced a new abstract class called `Focusable` to encapsulate the common `Optional` usage in `Actor` and `Attribute` classes. The `Focusable` class represents focusable objects and gives access and modification of a property or element in a given subject via Optional. Refactoring in the `Actor` and `Attribute` classes removes duplicate code, enhancing code readability and maintainability.