Update build script and add tsconfig.prod.json
The build script in package.json has been updated to use tsconfig.prod.json for TypeScript configurations. A new file, tsconfig.prod.json, has been introduced to define the TypeScript compilation settings for production.

The main reason for the split is that excluding the test dir seems to make the IDE's typechecker confused of certain signatures like that of World.get.